--- crates/builder/src/postbuild/mod.rs
@@ -1,5 +1,0 @@
-mod assembly;
-mod config;
-mod wasm;
-
-pub use config::PostbuildConfig;
