--- crates/builder/src/prebuild/mod.rs
@@ -1,12 +1,0 @@
-mod assembly;
-mod config;
-mod file;
-mod fontforge;
-mod localized;
-mod sass;
-
-pub use assembly::Assembly;
-pub use config::PrebuildConfig;
-pub use file::File;
-pub use localized::Localized;
-pub use sass::Sass;
